requirements.txt
# ============================================================
# Hallucination Detection System — Python Dependencies
# Project: Fine-Grained Hallucination Detection Using Hidden States
# Author: Chalani Dinitha (20211032)
# ============================================================

# ── LLM & Deep Learning ──────────────────────────────────────
torch==2.2.0
transformers==4.40.0
accelerate==0.29.0
sentencepiece==0.2.0
protobuf==4.25.3

# ── FastAPI Backend ───────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ── Signal Processing & Math ──────────────────────────────────
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.2

# ── NLP / NER for EAT Detection ──────────────────────────────
spacy==3.7.4

# ── Datasets & Evaluation ────────────────────────────────────
datasets==2.19.0
evaluate==0.4.1

# ── Experiment Tracking ───────────────────────────────────────
mlflow==2.12.1

# ── Testing ──────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# ── Utilities ────────────────────────────────────────────────
tqdm==4.66.2
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
jupyter==1.0.0
ipykernel==6.29.4
